requirements.txt
argon2-cffi==20.1.0
astroid==2.4.2
async-generator==1.10
attrs==20.2.0
backcall==0.2.0
bbox==0.9.2
bleach==3.2.1
cffi==1.14.3
click==7.1.2
colorlog==4.2.1
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
importlib-metadata==2.0.0
ipykernel==5.3.4
ipython==7.18.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==5.5.3
jedi==0.17.2
Jinja2==2.11.2
joblib==0.17.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.6.3
jupyterlab-pygments==0.1.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.1
nltk==3.5
notebook==6.1.4
numpy==1.19.2
opencv-python==4.4.0.44
packaging==20.4
pandas==1.1.3
pandocfilters==1.4.2
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
pkg-resources==0.0.0
prometheus-client==0.8.0
prompt-toolkit==3.0.8
ptyprocess==0.6.0
pycparser==2.20
Pygments==2.7.1
pylint==2.6.0
pyparsing==2.4.7
pyquaternion==0.9.5
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2020.1
pyzmq==19.0.2
qtconsole==4.7.7
QtPy==1.9.0
regex==2020.10.15
Send2Trash==1.5.0
six==1.15.0
terminado==0.9.1
testpath==0.4.4
toml==0.10.1
tornado==6.0.4
tqdm==4.50.2
traitlets==5.0.5
typed-ast==1.4.1
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
wrapt==1.12.1
zipp==3.3.1
